feat: setting up folder structure
