+++ src/App.tsx
@@ -1,9 +1,7 @@
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import { Header } from "./Header";
-import { Home } from "./Home";
-import { NotFound } from "./NotFound";
-import { Packages } from "./Packages";
+import { Header } from "./components";
 import { Theme } from "./Theme";
+import { Home, NotFound, Packages } from "./views";
 
 export function App() {
   return (
