+++ src/api/package/index.ts
@@ -1,0 +1,3 @@
+export { createAssembly } from "./assemblies";
+export { getAssetsPath, getFullPackageName } from "./util";
+export { fetchMetadata } from "./metadata";
