+++ src/api/package/metadata.ts
@@ -1,0 +1,3 @@
+export function fetchMetadata() {
+  return null;
+}
