+++ src/api/package/util.ts
@@ -8,15 +8,3 @@ export function getAssetsPath(name: string, version: string, scope?: string) {
 export function getFullPackageName(name: string, scope?: string) {
   return scope ? `${scope}/${name}` : name;
 }
-
-export function parseSearch(search: string) {
-  const params: any = {};
-  const keyValues = search.substring(1, search.length).split("&");
-
-  for (const kv of keyValues) {
-    const parts = kv.split("=");
-    params[parts[0]] = parts[1];
-  }
-
-  return params;
-}
