+++ src/components/GettingStarted/index.tsx
