+++ src/components/OperatorArea/index.tsx
@@ -1,0 +1,3 @@
+export function OperatorArea() {
+  return <div>Lets get some operator info amiright</div>;
+}
