+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { GettingStarted } from "./GettingStarted";
+export { Header } from "./Header";
+export { OperatorArea } from "./OperatorArea";
+export { PackageDetails } from "./PackageDetails";
+export { PackageDocs } from "./PackageDocs";
+export { PackageHeader } from "./PackageHeader";
+export { PackageLatest } from "./PackageLatest";
