+++ src/views/index.ts
@@ -1,0 +1,4 @@
+export { Home } from "./Home";
+export { NotFound } from "./NotFound";
+export { Package } from "./Package";
+export { Packages } from "./Packages";
